+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - updates to the `fi.json` language file (#2253) (@l1500s)
 - update default dashboard layout
 - add support for dark mode switching with custom plugin uis (@NorthernMan54)
+- Display Docker Version Details in the Update Information Box (#2488) (@NorthernMan54)
 
 ### Other Changes